add FastAPI implementation
The FastAPI version is intended to replace
the current swagger_server version.  It is almost
a drop-in replacement, though the authorization
is slightly different.

Changed the authorization client to work with
either version, by trying to use a form first.

Also make the ODIN_SCHEME an env var so
we can get around having to constantly pass
it.