1 rename client API to muninn, 2 use fastapi async
the client API in develop was odin.api but this caused
a lot of confusion in the package locations and proper installation.
rename the client package to muninn.

the implementation of the asyncio loop was ugly and
inefficient and seemed to cause run loop errors running
in uvicorn when the server is under stress.

now that we use fastapi, there is no reason for any of
this, just make any server calls that are using WS
async and await the websocket handler.