change async handling for fastAPI/uvicorn (#45)
update docs

WIP: update odin into 2 packages, server side and client (#43)

* update odin into 2 packages, server side and client

Update commands.md

Feature/add cpu limits/resources and security context (#46)

* k8s.py parse cpu limits/resources; parse security contetext

* k8s make fields under cpu and security_context optional

* test_pipeline.py add test for cpu & security_context

* add gitlab-ci.yml

* cleanup.py remove odin.api

* k8s.py check each attribut of security_context before setting

* k8s.py add V1PodSecurityContext to pod instead of container

* k8s.py clean up dict.get syntax

* remove .gitlab-ci

Feature/fastapi async (#44)

* try async direct

* 1 rename client API to muninn, 2 use fastapi async

the client API in develop was odin.api but this caused
a lot of confusion in the package locations and proper installation.
rename the client package to muninn.

the implementation of the asyncio loop was ugly and
inefficient and seemed to cause run loop errors running
in uvicorn when the server is under stress.

now that we use fastapi, there is no reason for any of
this, just make any server calls that are using WS
async and await the websocket handler.

bump version

Feature/docs and tests (#47)

* pipelines.md update docs for pipeline, notes on security context & cpu resource

* core.py read_pipeline_config takes predefined pipeline id instead of always generate random

* add task_to_pod_spec test

* task_to_pod_specs test update reference

bump

try again

add odin-api

not sure how this happened

Hotfix/fix pkg (#50)

* find

* bump version