add WS health subprop
For now, if a WS conn is made from /app, itll
return with a status of "responsive" for
"ws_server_status" field.  This might not be ideal:
if ws server is down, youll get a 500, and if its up
but slow, this call will block for a long time.

probably neither is idea, so might want to change
the approach a bit